add playable + clear player prefb menu
