+++ SCripts/Editor/MenuEditor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 85dffac7440c04bb880b5bd667f2e62a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
